requirements.txt
Flask==2.3.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
PyYAML==6.0.1
Werkzeug==3.0.1
